requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.0.0
celery>=5.3.0
redis>=5.0.0
PyJWT>=2.8.0
pandas>=2.0.0
openpyxl>=3.1.0
pytest>=7.4.0
pytest-django>=4.5.0
factory-boy>=3.3.0
coverage>=7.3.0
Faker>=20.0.0
PyInstaller>=6.0.0
django-celery-beat>=2.5.0
